Replace 32bit IPv6 accessors with 16bit ones when comparing addresses.
Also added simple unit tests to check the comparisons are correct.

+++ src/common/testlib.h
@@ -98,7 +98,7 @@ int delay_send_packet(int sock, char *packet, int size, struct addrinfo *dest,
         uint32_t inter_packet_delay, struct timeval *sent);
 char *address_to_name(struct addrinfo *address);
 int compare_addresses(const struct sockaddr *a,
-        const struct sockaddr *b, int len);
+        const struct sockaddr *b, uint8_t len);
 struct addrinfo *get_numeric_address(char *interface, char *port);
 int bind_socket_to_device(int sock, char *device);
 int bind_sockets_to_device(struct socket_t *sockets, char *device);
